[fix] Fixed several occasional crashes And disable unreasonable 'Exit' menus for Android.

+++ frontend/ui/quickstart.lua
@@ -85,12 +85,14 @@ function QuickStart:getQuickStart()
     local quickstart_filename = ("%s/quickstart-%s-%s.html"):format(quickstart_dir, language, rev)
     if lfs.attributes(quickstart_filename, "mode") ~= "file" then
         -- purge old quickstart guides
-        local iter, dir_obj = lfs.dir(quickstart_dir)
-        for f in iter, dir_obj do
-            if f:match("quickstart-.*%.html") then
-                local file_abs_path = FFIUtil.realpath(("%s/%s"):format(quickstart_dir, f))
-                os.remove(file_abs_path)
-                DocSettings:open(file_abs_path):purge()
+        local ok, iter, dir_obj = pcall(lfs.dir, quickstart_dir)
+        if ok then
+            for f in iter, dir_obj do
+                if f:match("quickstart-.*%.html") then
+                    local file_abs_path = FFIUtil.realpath(("%s/%s"):format(quickstart_dir, f))
+                    os.remove(file_abs_path)
+                    DocSettings:open(file_abs_path):purge()
+                end
             end
         end
